updated nic_utilities

+++ boards/v0/testing/nic_apis/nic_utilities.h
@@ -33,18 +33,18 @@ void writeNicReg(uint32_t regIndex, uint32_t value);
 void nicRegInit(CortosQueueHeader* Free_Queue, CortosQueueHeader* Rx_Queue,CortosQueueHeader* Tx_Queue);
 
 // to load Ethernet header
-// 		eth_hdr should be array of length 2 and width 64.
+// 		eth_hdr should be array of length 4 and width 32.
 //		ethernet header is written to eth_hdr array.
-void loadEthernetHeader(uint32_t *buffer_addr, uint64_t *eth_hdr);
+void loadEthernetHeader(uint32_t *buffer_addr, uint32_t *eth_hdr);
 
 // to load payload
 void loadPayload(uint32_t *buffer_addr , uint64_t *payload);
 
 // to read control data of the buffer
-void readBufControlData(uint32_t *buffer_addr, uint16_t *packetSize, uint16_t *lastTkeep);
+void readBufControlData(uint32_t *buffer_addr, uint16_t *packetSize, uint8_t *lastTkeep);
 
 // to modify control data of the buffer
-void writeBufControlData(uint32_t *buffer_addr, uint16_t packetSize, uint16_t lastTkeep);
+void writeBufControlData(uint32_t *buffer_addr, uint16_t packetSize, uint8_t lastTkeep);
 
 
 #endif
